Improve workflow for development
This commit adds [yalc](https://github.com/wclr/yalc) as a dev dependency. Yalc creates a local store for publishing packages that you can then link as dependencies in other projects. The build step now creates a `.yalc/` directory in the project folder with the latest build. This can be mounted by `check-web` and other projects in their docker-container volumes in their node_modules directory. The yalc command line tool can be used in their build processes to link to the new volume, watch the directory to see if it changes, etc.