MetadataFile selector now supports click-to-upload
Previously, only drag and drop worked on the file upload box, even though the user is also told they could click the box to upload a file. This updates the dropZone area so that it is now clickable (via an invisible `<input/>` element stored on the react component).

This abstracts out the validation (file too big, wrong file type) so both drag & drop and click functions use the same validation, and the CSS now correctly turns the cursor into a "hand" style pointer to indicate to the user that the area is clickable like a button.