v0.1.43 - parse dirty date strings
There are about a dozen date string formats in our production database. Two of them are standard. Three more nonstandard formats seem to cover 99.9% of cases, with 12 records that are nearly garbage. This commit lets us parse those 99.9% of records and should fail out gracefully on the remaining 12 (displaying a wrong time zone offset or time, but not crashing).

Fixes CHECK-1354.